Move the IoTDBTimeZoneIT to integration-test (#6427)
